backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-community
sentence-transformers
faiss-cpu
numpy
pytesseract
pdfplumber
python-docx
PyPDF2
fastapi
uvicorn
python-multipart
pydantic
pydantic[email]
sentence-transformers
faiss-cpu
pypdf2
python-docx
markdown
beautifulsoup4
python-dotenv
requests
